Finalized
Added a sync(2) call to the end of recover/cleanup just to ensure that
everything is on disk. Also created a README which has some basic info
such as how to compile and run. Also added the image that will be copied
to the drive and then deleted during the demo.